feat: create text component
