+++ src/components/atoms/Text/index.tsx
@@ -1,21 +1,11 @@
 import React from 'react';
-import { StyleProp, TextStyle } from 'react-native';
 import * as S from './styles';
-
-export type Color = 'primary' | 'secondary' | 'black' | 'white';
-
-export type CustomTextProps = {
-  family?: 'regular' | 'bold';
-  color?: Color;
-  size?: 'ultra_small' | 'small' | 'medium' | 'big' | 'large';
-  children: string;
-  style?: StyleProp<TextStyle>;
-};
+import { CustomTextProps } from './types';
 
 const Text: React.FC<CustomTextProps> = ({
   children,
   family = 'regular',
-  color = 'black',
+  color = 'primary',
   size = 'medium',
   style,
   ...props
